requirements.txt
future
albumentations
matplotlib
numpy
opencv-python
pandas
Pillow
pyrr
pytest
ruamel.yaml
seaborn
scipy
tensorboardX
torch
torchvision
tqdm
webcolors
gdown
